requirements.txt
numpy==2.4.1
paho-mqtt==1.6.1
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.3
